+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -8,6 +8,7 @@ Release History
 
 2.0.25
 ++++++
+* msi: output principal id on enabling system assigned identity
 * vm image: support accept market terms to use vm images
 * vm/vmss create: ensure commands can run under proxy with unsigned certificates.
 * vmss:(PREVIEW) support low priority